remove deprecated etls and rename data_model to cris_import
