--- atd-etl/cr3_download/.dockerignore
@@ -1,2 +1,0 @@
-env
-*.env
