--- atd-etl/cr3_download/env_template
@@ -1,3 +1,0 @@
-OP_API_TOKEN=
-OP_CONNECT=
-OP_VAULT_ID=
