--- atd-etl/cr3_download/requirements.txt
@@ -1,5 +1,0 @@
-awscli==1.*
-boto3==1.*
-python-magic==0.*
-requests==2.*
-onepasswordconnectsdk==1.*
