--- atd-etl/cr3_extract_diagram/.dockerignore
@@ -1,2 +1,0 @@
-env
-*.env
