--- atd-etl/cr3_extract_diagram/env_template
@@ -1,5 +1,0 @@
-HASURA_ENDPOINT=
-HASURA_ADMIN_KEY=
-AWS_DEFAULT_REGION=
-AWS_ACCESS_KEY_ID=
-AWS_SECRET_ACCESS_KEY=
