+++ atd-etl/cris_import/requirements.txt
@@ -1,7 +1,3 @@
-onepasswordconnectsdk==1.*
-sysrsync==1.*
-psycopg2-binary==2.*
-sshtunnel==0.4.*
-requests==2.*
 boto3==1.*
-python-magic==0.*
+pdf2image==1.*
+requests==2.*
