--- atd-etl/data_model/.dockerignore
@@ -1,4 +1,0 @@
-env
-*.env
-**.pyc
-extracts/*
