--- atd-etl/data_model/.gitignore
@@ -1,2 +1,0 @@
-extracts/*
-!extracts/*.md
