--- atd-etl/data_model/requirements.txt
@@ -1,3 +1,0 @@
-boto3==1.*
-pdf2image==1.*
-requests==2.*
