--- atd-etl/populate_cr3_file_metadata/.dockerignore
@@ -1,1 +1,0 @@
-env
