--- atd-etl/populate_cr3_file_metadata/.gitignore
@@ -1,3 +1,0 @@
-*pdf
-env
-
