--- atd-etl/populate_cr3_file_metadata/docker-compose.yaml
@@ -1,10 +1,0 @@
-services:
-  metadata:
-    build:
-      context: .
-      dockerfile: Dockerfile
-    volumes:
-      - .:/app
-    entrypoint: /bin/bash
-    env_file:
-      - env
