--- atd-etl/populate_cr3_file_metadata/env_template
@@ -1,7 +1,0 @@
-PDF_MAX_RECORDS=100
-AWS_BUCKET_NAME=
-AWS_BUCKET_ENVIRONMENT=
-HASURA_ENDPOINT=
-HASURA_ADMIN_KEY=
-AWS_ACCESS_KEY_ID=
-AWS_SECRET_ACCESS_KEY=
