--- atd-etl/populate_cr3_file_metadata/requirements.txt
@@ -1,3 +1,0 @@
-boto3==1.33.*
-requests==2.31.*
-tqdm==4.*
